Delete unused files and functions, change function file names, add explanations to some functions, delete debugging printout
