--- include/cylinder.h
@@ -1,17 +1,0 @@
-//
-// Created by 宋飞龙 on 2018-11-11.
-//
-
-#ifndef INTRODUCTION_CYLINDER_H
-#define INTRODUCTION_CYLINDER_H
-
-#include <Eigen/Core>
-
-void cylinder(
-      int div,
-      int part,
-      Eigen::MatrixXd & LV,
-      Eigen::MatrixXi & LF
-    );
-
-#endif //INTRODUCTION_CYLINDER_H
