+++ src/triangle_hunt.cpp
@@ -9,42 +9,8 @@
 
 using namespace std;
 
-void triangle_hunt(Eigen::Matrix3d & P, Eigen::MatrixXd & V, Eigen::MatrixXi & F) {
+int triangle_hunt(Eigen::Matrix3d & P, Eigen::MatrixXd V, Eigen::MatrixXi F) {
     Eigen::RowVector3d triangleCenter = (P.row(0) + P.row(1) + P.row(2)) / 3.0;
-    double dist = std::numeric_limits<double>::infinity();
-    int fi = -1;
-
-    for (int i = 0; i < F.rows(); i++) {
-        Eigen::RowVector3i f = F.row(i);
-        Eigen::RowVector3d p0 = V.row(f(0));
-        Eigen::RowVector3d p1 = V.row(f(1));
-        Eigen::RowVector3d p2 = V.row(f(2));
-
-        Eigen::RowVector3d pc = (p0 + p1 + p2) / 3.0;
-
-        double cdist = abs((pc - triangleCenter).norm());
-
-        if (cdist < dist) {
-            dist = cdist;
-            fi = i;
-        }
-    }
-
-    // Change the location
-    Eigen::RowVector3i f = F.row(fi);
-    cout<<"triangle hunt V"<<endl;
-    cout<<V.row(f(0))<<endl;
-    V.row(f(0)) = P.row(0);
-    V.row(f(1)) = P.row(1);
-    V.row(f(2)) = P.row(2);
-    cout<<"triangle hunt P"<<endl;
-    cout<<P.row(0)<<endl;
-}
-
-int triangle_hunts(Eigen::Matrix3d & P, Eigen::MatrixXd V, Eigen::MatrixXi F) {
-    Eigen::RowVector3d triangleCenter = (P.row(0) + P.row(1) + P.row(2)) / 3.0;
-    std::cout << "triangle center" << std::endl;
-    std::cout << triangleCenter << std::endl;
     double dist = std::numeric_limits<double>::infinity();
     int fi = -1;
     Eigen::RowVector3d sp0 = P.row(2);
@@ -53,8 +19,6 @@ int triangle_hunts(Eigen::Matrix3d & P, Eigen::MatrixXd V, Eigen::MatrixXi F) {
 
     for (int i = 0; i < F.rows(); i++) {
         Eigen::RowVector3i f = F.row(i);
-
-//        Eigen::RowVector3d pc = (p0 + p1 + p2) / 3.0;
 
         double minlensum = std::numeric_limits<double>::infinity();
         for (int k = 0; k < 3; k++) {
@@ -69,7 +33,6 @@ int triangle_hunts(Eigen::Matrix3d & P, Eigen::MatrixXd V, Eigen::MatrixXi F) {
             }
         }
 
-//        double cdist = abs((pc - triangleCenter).norm());
         double cdist = minlensum;
 
         if (cdist < dist) {
